🚚 move : team.handler.js -> chatCommand.handler.js , 🔥 del : team.handler.js 삭제 , ♻️ refactor : chatCommand.handler.js 와 chatCommands.js 수정
